+++ src/Position/Positionable.php
@@ -40,6 +40,8 @@ interface Positionable
      *        Providing zero for this parameter is deprecated in version 8.9.0, and will not be supported from v9.0.
      *        Use `null` instead when no line number is available.
      * @param int<0, max>|null $columnNumber
+     *
+     * @return $this fluent interface
      */
-    public function setPosition(?int $lineNumber, ?int $columnNumber = null): void;
+    public function setPosition(?int $lineNumber, ?int $columnNumber = null): Positionable;
 }